added parsing for set statement (only integer values so far); added code generator

+++ components/interpreter/types.hpp
@@ -7,9 +7,9 @@ namespace Interpreter
 
     typedef unsigned int Type_Data; // 32 bit
     
-    typedef unsigned int Type_Integer; // 32 bit
+    typedef int Type_Integer; // 32 bit
     
-    typedef unsigned int Type_Float; // 32 bit
+    typedef float Type_Float; // 32 bit
 }
 
 #endif
